+++ js/plugins/splitter.js
@@ -50,7 +50,7 @@ var Splitter = {
         var that = this, element = this.element, o = this.options;
         var children = element.children(o.children).addClass("split-block");
         var i, children_sizes = [];
-        var gutters;
+        var gutters, resizeProp = o.splitMode === "horizontal" ? "width" : "height";
 
         if (!Utils.isValue(element.attr("id"))) {
             element.attr("id", Utils.elementId("splitter"));
@@ -62,9 +62,7 @@ var Splitter = {
         }
 
         for (i = 0; i < children.length - 1; i++) {
-            $("<div>").addClass("gutter").css({
-                flexBasis: o.gutterSize
-            }).insertAfter($(children[i]));
+            $("<div>").addClass("gutter").css(resizeProp, o.gutterSize).insertAfter($(children[i]));
         }
 
         gutters = element.children(".gutter");
@@ -90,7 +88,7 @@ var Splitter = {
                 }
             } else {
                 $.each(children, function(){
-                    this.style.setProperty('min-height', String(o.minSizes).contains("%") ? o.minSizes : String(o.minSizes).replace("px", "")+"px", 'important');
+                    this.style.setProperty('min-'+resizeProp, String(o.minSizes).contains("%") ? o.minSizes : String(o.minSizes).replace("px", "")+"px", 'important');
                 });
             }
         }
@@ -115,7 +113,7 @@ var Splitter = {
 
             $(window).on(Metro.events.move + "-" + element.attr("id"), function(e){
                 var pos = Utils.getCursorPosition(element, e);
-                var new_pos, func = o.splitMode === "horizontal" ? 'outerWidth' : 'outerHeight';
+                var new_pos;
 
                 if (o.splitMode === "horizontal") {
                     new_pos = (pos.x * 100 / w) - (start_pos.x * 100 / w);