+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+### 4.2.27
+
 ### 4.2.26
 + [x] Image compare: fix for touch devices
 + [x] Image magnifier: fix for touch devices